split debug_traceTransaction scenario to 3 different ones
split debug_traceTransaction scenario to 3 different ones
