+++ loadtesting/k6/scenarios/tracerStandardLoad.js
@@ -1,4 +1,6 @@
-import { default as DebugTraceTransactionTest } from '../tests/tracer/debugTraceTransaction.test.js'
+import { default as DebugTraceTransactionTest } from '../tests/tracer/traceTransaction/debugTraceTransactionSimple.test.js'
+import { default as DebugTraceTransactionWithEventsTest } from '../tests/tracer/traceTransaction/debugTraceTransactionWithEvents.test.js'
+import { default as DebugTraceIterativeTransactionTest } from '../tests/tracer/traceTransaction/debugTraceTransactionIterative.test.js'
 import { default as EthCallTest } from '../tests/tracer/ethCall.test.js';
 import { default as EthGetBalanceTest } from '../tests/tracer/ethGetBalance.test.js';
 import { default as EthGetTransactionCountTest } from '../tests/tracer/ethGetTransactionCount.test.js';
@@ -9,6 +11,26 @@ export const options = {
     scenarios: {
         DebugTraceTransaction: {
             exec: 'DebugTraceTransaction',
+            executor: 'ramping-vus',
+            startVUs: 0,
+            stages: [
+                { duration: '20s', target: usersNumber },
+                { duration: '20m', target: usersNumber },
+            ],
+            gracefulRampDown: '60s',
+        },
+        DebugTraceTransactionWithEvents: {
+            exec: 'DebugTraceTransactionWithEvents',
+            executor: 'ramping-vus',
+            startVUs: 0,
+            stages: [
+                { duration: '20s', target: usersNumber },
+                { duration: '20m', target: usersNumber },
+            ],
+            gracefulRampDown: '60s',
+        },
+        DebugTraceIterativeTransaction: {
+            exec: 'DebugTraceIterativeTransaction',
             executor: 'ramping-vus',
             startVUs: 0,
             stages: [
@@ -63,6 +85,14 @@ export const options = {
 
 export function DebugTraceTransaction() {
     DebugTraceTransactionTest();
+}
+
+export function DebugTraceTransactionWithEvents() {
+    DebugTraceTransactionWithEventsTest();
+}
+
+export function DebugTraceIterativeTransaction() {
+    DebugTraceIterativeTransactionTest();
 }
 
 export function EthCall() {
